Изменил тип данных в моделях с Double на BigDecimal для большей точности

+++ Homework/task7/src/main/java/org/yakdanol/task7/model/ConversionResponse.java
@@ -3,6 +3,8 @@ package org.yakdanol.task7.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class ConversionResponse {
@@ -11,5 +13,5 @@ public class ConversionResponse {
 
     private String toCurrency;
 
-    private double convertedAmount;
+    private BigDecimal convertedAmount;
 }
